two fixes for git clone of xc-vusb:
1. $false and $true were spelled false and true, which broke the logic
2. There were a couple of errors in the git invocations.

Essentially and embarassingly the code path taken on first use of this
was untested in the version I sent for review before. But this time I
have tested this path. I need a better way to set the git cloning URLs
to my own so that I safely delete everything and do a clean rebuild.