Fix xenbus to cause surprise removals for vusb device on removal.
The Windows front end vusb driver expects to be removed via a
surprise removal. The existing code was trying to do an eject
which was not causing the front end driver to finish cleaning up
the device that was removed. Custom code for surprise removal was
only added for vusb devices to prevent injuring the already delicate
shutdown/removal code for other types of devices.

OXT-35

Signed-off-by: Ross Philipson ross.philipson@gmail.com