Fix xenbus to not re-enumerate closed VUSB devices.
When VUSB devices are removed, they are just gone for good. There
is no possibility of a reconnect. This fix detects the backend state
went to closed which means the device was there and gone. In this
case, xenbus just ignores the closed device. The usbowls thingy can
then properly clean up the mess in xenstore.

OXT-35

Signed-off-by: Ross Philipson ross.philipson@gmail.com